feat(desktop): replace parseRuntimeEnv with t3-env layer, drop sparklePublicKey
Phase B of 2026-04-23-desktop-pre-release-batch. Introduces
src/env/main.ts (main process, isServer: true) and src/env/renderer.ts
(VITE_* prefix, import.meta.env). Deletes src/shared/env.ts (silent-fail
parseRuntimeEnv); rehomes buildInfoSchema to src/shared/build-info-schema.ts.

Sentry DSN is now baked into the main bundle via a __SENTRY_DSN__ custom
token define in vite.main.config.ts (per Sentry's Electron guidance —
avoids replacing process.* keys). mainEnv reads the token with a
process.env.SENTRY_DSN fallback for dev.

Collapses the two duplicated getApiOrigin() helpers in index.ts and
auth-flow.ts into mainEnv.LIGHTFAST_API_URL with a
.default("http://localhost:3024").

forge.config.ts: hardcode publisher repo to lightfastai/lightfast; remove
LIGHTFAST_DESKTOP_RELEASE_REPO indirection.

README: drop LIGHTFAST_DESKTOP_RELEASE_REPO dry-run reference; point
contributors at .env.example (added in Phase F).
