+++ apps/desktop/package.json
@@ -25,6 +25,7 @@
     "@electron/notarize": "^3.1.1",
     "@electron/osx-sign": "^1.3.3",
     "@repo/typescript-config": "workspace:*",
+    "@t3-oss/env-core": "catalog:",
     "@types/electron-squirrel-startup": "^1.0.2",
     "@types/node": "catalog:",
     "@types/react": "catalog:react19",
@@ -60,6 +61,5 @@
   },
   "buildFlavor": "dev",
   "buildNumber": "1",
-  "sparkleFeedUrl": "",
-  "sparklePublicKey": ""
+  "sparkleFeedUrl": ""
 }
