+++ apps/desktop/src/main/bootstrap.ts
@@ -1,6 +1,7 @@
 import { join } from "node:path";
 import { app } from "electron";
 import squirrelStartup from "electron-squirrel-startup";
+import { mainEnv } from "../env/main";
 
 if (squirrelStartup) {
   app.quit();
@@ -11,18 +12,11 @@ app.setName(productName);
 app.setPath("userData", join(app.getPath("appData"), productName));
 
 if (!app.isPackaged) {
-  const raw = process.env.LIGHTFAST_REMOTE_DEBUG_PORT?.trim();
-  if (raw) {
-    const port = Number(raw);
-    if (Number.isInteger(port) && port >= 1 && port <= 65_535) {
-      app.commandLine.appendSwitch("remote-debugging-port", String(port));
-      app.commandLine.appendSwitch("remote-debugging-address", "127.0.0.1");
-      console.log(`[cdp] remote debugging on 127.0.0.1:${port}`);
-    } else {
-      console.warn(
-        `[cdp] ignoring LIGHTFAST_REMOTE_DEBUG_PORT=${raw}; expected integer 1-65535`
-      );
-    }
+  const port = mainEnv.LIGHTFAST_REMOTE_DEBUG_PORT;
+  if (port !== undefined) {
+    app.commandLine.appendSwitch("remote-debugging-port", String(port));
+    app.commandLine.appendSwitch("remote-debugging-address", "127.0.0.1");
+    console.log(`[cdp] remote debugging on 127.0.0.1:${port}`);
   }
 }
