+++ apps/desktop/src/shared/ipc.ts
@@ -49,7 +49,6 @@ export interface BuildInfoSnapshot {
   buildNumber: string;
   name: string;
   sparkleFeedUrl: string;
-  sparklePublicKey: string;
   version: string;
 }
