+++ apps/desktop/tsconfig.json
@@ -7,6 +7,7 @@
     "moduleResolution": "Bundler"
   },
   "include": [
+    "src/env/renderer.ts",
     "src/renderer/**/*.ts",
     "src/renderer/**/*.tsx",
     "src/shared/**/*.ts"
