+++ apps/desktop/tsconfig.node.json
@@ -6,6 +6,7 @@
     "moduleResolution": "Bundler"
   },
   "include": [
+    "src/env/main.ts",
     "src/main/**/*.ts",
     "src/preload/**/*.ts",
     "src/shared/**/*.ts",
