+++ apps/desktop/vite.main.config.ts
@@ -1,6 +1,9 @@
 import { defineConfig } from "vite";
 
 export default defineConfig({
+  define: {
+    __SENTRY_DSN__: JSON.stringify(process.env.SENTRY_DSN ?? ""),
+  },
   build: {
     lib: {
       entry: "src/main/bootstrap.ts",
